requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-decouple>=3.8
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
django-extensions>=3.2.0
